pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ai-video-editor"
version = "0.1.0"
description = "AI-assisted transcript-based video editing pipeline"
requires-python = ">=3.13"
dependencies = [
    "typer>=0.12",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "loguru>=0.7",
    "ffmpeg-python>=0.2",
    "noisereduce>=3.0",
    "soundfile>=0.12",
    "numpy>=1.26",
    "whisperx>=3.8",
    "python-dotenv>=1.0",
    "langchain-google-genai>=4.0",
    "langgraph>=0.4",
    "elevenlabs>=2.40",
    "rapidfuzz>=3.0",
    "sentence-transformers>=5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]

[project.scripts]
ai-video-editor = "ai_video_editor.cli.app:main"

[tool.hatch.build.targets.wheel]
packages = ["ai_video_editor"]
